requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
awscli==1.32.98
bcrypt==4.1.3
blinker==1.8.1
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.16
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pathlib==1.0.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
stripe==9.5.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
